Avoid race condition
I don't understand it well, but this assertion is failing sometimes in my
computer. Maybe it's because Cypress doesn't not respect the `delay`
option when the returned status is 503, even though there is another
test previously that should fail if that would be true.

I'll remove it from now as it doesn't seem that important.